Ignore Ruby version specified in Gemfile
Ignore any explicit Ruby version specified in a Gemfile, as a mismatch with the
system Ruby version that Bump is running on causes an error during dependency
resolution.

Ideally we would run the DependencyFileUpdaters::Ruby class using whichever
Ruby version was specified in the Gemfile (or in the .ruby_version file, which
we're not even fetching at the moment...), but that's impractical, and it's
better to produce a PR for the user with gems that require a bump to their
Ruby version (and will blow up in CI) than not to produce a PR at all.