Do JavaScript file parsing in JavaScript
Adds JavaScript helpers that use yarn as a library to parse the
yarn.lock. This means we now extract the correct versions of JavaScript
dependencies, as we're relying on the lockfile rather than the
package.json.

This also introduces the concept of language helpers, which will be used
more in JavaScript, and possibly for other languages in due course.