Manually set Bundler root
When the bundler root is not set to the working directory, it attempts to
add the full expanded path to the Gemfile for gems that are path based.
This resulted in an erroneous path being added to the updated Gemfile
due to the temporary working directory not being on the original
repository.