NPM: Handle 403 Forbidden when updating lockfiles
Seems to handle the case when we auth against artifactory and get
"Unable to authenticate, need: Basic realm="Artifactory Realm""

This seems to be caused by a trailing slash in the registry host as set
in config variables. When we write this out in npmrc we add another
trailing slash and this seems to cause the above error. Changing to a
single trailing slash returns a 403 Forbidden error.