go_modules: remove unnecessary reqs from go.mod
Reorganizes a bit of code (class constants, attr_reader), which creates
some visual noise in the diff. Sorry about that.

Previously, we'd:
1. Add the new dependency versions to the go.mod
2. If a go.sum was present, run `go get` to update the go.sum

Sometimes the new dependency requirement should change other entries in
the go.mod. For instance, say we use dependency x@v1, which depends on
y@v1. We upgrade y to v2 manually in the go.mod, which promotes y@v2 to
the go.mod. Dependabot kicks in and then updates x to v2, which includes
a dependency on y@v2, meaning the go.mod entry for y@v2 is no longer
necessary (according to MVS).

With the previous algorithm, we wouldn't update the go.mod
correspondingly.

Now, we:
1. Add the new dependency versions to the go.mod
2. Run `go get` to produce a regenerated go.sum and go.mod
3. Figure out all the differences in `require`s between the new go.mod
   and the input go.mod
4. Update the original go.mod to reflect all these changes