Remove support for jinja requimrents files
The functionality that this relied on had been deprecated for a long
time, and has since been removed from pip.

The featrure was removed in: https://github.com/pypa/pip/pull/8098

I don't see a clean way for us to support this, the only option would be
to patch pip or to maintain a custom fork, which I don't think is worth
it as it will significantly impact our ability to keep the dependency up
to date and for Dependabot to work the same way as pip does.