Update python to work with requirements of form `file:.`
Currently if a requirement file contains a path dependency where the
path starts with `file:`, the file fetcher will try and find a file
called `/file:./setup.py`, which fails because it should be looking for
`setup.py`.

This commit updates Dependabot::Python::FileFetcher to ignore the
`file:` part if it is present.