[WIP] Go mod tidy support
This PR runs `go mod tidy` after updating go modules.

In order to do this, we rely on the full repo being checked out, and
because we now have access to the full source code, the implementation
has changed from a series of changes on temporary go modules, to running
the update in the context of the actual repo.

In order to properly test this, the test setup has changed. Instead of
updating the fixtures in tests, we set up a project folder for each
test, that mimics a minimal repository. This project is then copied over
and turned into a repository in the tests, and we perform the updates on
this repository.