Allow requirements.txt files of up to 200kb
We have a limit on the size of .txt files that we fetch for Python
repositories, most likely because we try to fetch _all_ files with that
extension, and the GitHub contents API only allows fetching files up to
1mb.

Some users are now running into the 100kb size limit that we've set, and
are getting back a pretty opaque `DependencyFileNotFound` error.

In order to resolve the issue for the cases that have been reported,
increasing to 200kb seems like a reasonable first step.