Update tests and fixtures for new Cargo.lock format
In https://github.com/rust-lang/cargo/pull/7070 the lockfile format was
updated, instead of a top-level `[metadata]` key that holds the
checksums for dependencies, these checksums are now specified in-line
with the dependency.

Since we shell out to `cargo` for most things, and in our tests only
explicitly check for the presence of the `sha` values, this change is
fairly minimal.

I've opted to remove the checksums in most of our fixtures, as they are
not strictly needed in our tests. I could have left them in there and
things would still pass (`cargo` will silently upgrade them to the new
format), however it seemed noisy.