Add lowest_security_fix_version to update checkers
In order to better explain why a security update is not possible, we're
exposing the `lowest_security_fix_version` which returns the lowest
available non-vulnerable version from the package registry not taking
resolvability into account. We'll use this alongside the existing method
`lowest_resolvable_security_fix_version` to explain which version is the
latest possible version that can be installed and which version is the
earliest non-vulnerable/fixed version.

Co-authored-by: Philip Harrison <philip@mailharrison.com>