Stop parsing unrequired indirect go packages
Previously we would run `go list -m -json all` to generate a list of
dependencies to attempt to update.

The way that this is implemented, it causes _every_ transitive
dependency to be listed, even if they are not required by this project.

This can happen when a module defines several packages, but only a
subset of those packages is used by the current repository.

It will also verify that each of those transitive dependencies is
available, and this causes issues when we rely on a dependency with
private transitive dependencies. We run into the scenario where we have
access to any of the transitive dependencies that are used, but since we
do not have access to an _unused_ transitive dependency, we run into a
`dependency_file_not_resolvable` error.

This is especially unfortunate, since we won't attempt to update _any_
indirect dependencies for go_modules anyway.

By using `go mod edit -json` instead, we get an easier to parse list of
packages that are _actually_ used in this repo.

In order to replicate the previous behavior of raising a
`DependencyFileNotResolvable` error when some of the packages are not
available, we now also run a `go get`, but it's debatable if we even
need to do this, since this error will be raised in subsequent steps (in
the `FileUpdater`), and not raising it here means that we can
consolidate our error handling for this scenario in one place.

Co-authored-by: Fatih Arslan <fatih@github.com>