Verify if composer name is compatible with v2
Composer v2 introduced a check on package names, and when no lockfile is
present to validate the plugin-api-version this causes us to attempt the
update with v2, which will fail because of this invalid name.

This includes an extra check to make sure the name is compatible with
the constraints of composer v2.