Include native helpers in bundler gem
When we moved this code out into native helpers, those were not included
in the gem that we build.

This has historically not really been an issue for us, since we run
Dependabot in the Docker environment where these helpers are manually
copied over. But for people using the gem as a standalone, this means it
no longer works.

Fixes #2885