Sanitize mentions more aggressively
The current sanitization works well when directly creating PRs, however,
when a user replies to such an email, some of the markup of the original
message will be stripped out (either by GitHub, or by email providers),
which can result in a link like `<a href="https://github.com/jurre>@jurre</a>`
being turned into just `@jurre`, which will consequently mention the
user when that email reply is rendered.

This wraps the inner mention in a codeblock, which should prevent this
from happening. It looks _slightly_ funny:

<a href="https://github.com/jurre><code>@jurre</code></a>

But I think this is worth it, getting unexpected mentions by Dependabot
is a source of frustration for users, and by showing that the mention is
wrapped in a codeblock, this should make it easier to spot if it was an
actual mention or not.

Alternatively we could insert some no-space character between the `@`
and the user login, but I personally prefer this more explicit approach.