Nuget: Parse floating notation when used in range
Typically you’d define a Nuget range as `[2.0, 3.0)` which translates to: "higher than
or equal to 2.0 but lower than 3.0". You can also say `2.0.*` which means any
patch version of 2.0.

However, it's valid to combine these like `[2.1.*,3.0.0)`.

The floating notation (`*`) seems like it’s redundant in this case,
as the `3.0.0)` bit would allow 2.2 to be installed.

I could not find any documentation on this, so I made a best guess
effort to do what seems right, previously a notation like this would
fail and cause the update to error. I've verified that this resolves
those errors when dry-running.