Don't raise when adding temp platform extensions
It's not clear to me why we need to raise in this scenario, those errors
are not handled. It seems better to attempt the update without adding
the extensions, for some cases that I've tried locally that seems to
update dependencies just fine.

Maybe this behavior was improved with composer 2?