Dependabot::UnexpectedExternalCode
This introduces the above new error, to be raised when Dependabot
determines it may need to evaluate external code to update a dependency.

The included sample and test cases centre around the Bundler ecosystem,
where Git sources which may result in a evaluating the dependency's
`.gemspec`, e.g. https://github.com/dependabot/dependabot-core/blob/main/common/dependabot-common.gemspec

Note: this guard and the associated exception are an opt-in feature, to
be used in conjunction with features that extend the resources
accessible to Dependabot (e.g. private packages and sources).