Insert zero-width space in @mentions
Insert zero-width space in @mentions to prevent sending notifications to
users tagged in changelogs and release notes when email-replying to pull
requests from dependabot.

When sending an email reply to a pull request from dependabot the
contents of the pull request description including changelogs can be
included in the body of the email causing notifications to be sent to
the @mentioned users.

We've previously attempted to fix this by changing @mentions into code
blocks but it seems like the email content is treated as plain text
removing any formatting before detecting usernames.

Ref: https://github.com/dependabot/dependabot-core/pull/2982