Remove go dep entirely
This has been deprecated for a while, both here in dependabot and also
in the `go` ecosystem. The ecosystem has moved to `go mod` (which we
already support).

So remove the `go dep` code entirely. No need to keep supporting it.

Anyone who wants this old of code is likely not keeping up with their
dependencies anyway, so probably will just keep using an old version of
`dependabot`.

See also https://github.com/dependabot/dependabot-core/issues/3529.