Terraform: Do not set dependency.version for version ranges
When a dependency specifies a version range, we do not want to set the
dependency objects version, as we cannot reliably determine it. Instead,
we need to rely on the _requirement_. For example, `~> 2.0` is not a
valid _version_, so when parsing it, we should set the `version`
attribute to nil.

This was already handled for terraform modules, but not for providers.